UP-2832 Don't remove guest layout from cache
git-svn-id: https://source.jasig.org/uPortal/branches/rel-3-1-patches@21848 f5dbab47-78f9-eb45-b975-e544023573eb

+++ uportal-impl/src/main/java/org/jasig/portal/layout/dlm/DistributedLayoutManager.java
@@ -149,8 +149,10 @@ IFolderLocalNameResolver
         final ApplicationContext applicationContext = PortalApplicationContextLocator.getApplicationContext();
         this.layoutCachingService = (ILayoutCachingService)applicationContext.getBean("layoutCachingService", ILayoutCachingService.class);
         
-        // Ensure a new layout gets loaded whenever a user logs in...
-        this.layoutCachingService.removeCachedLayout(owner, profile);
+        // Ensure a new layout gets loaded whenever a user logs in except for guest users
+        if (!owner.isGuest()) {
+            this.layoutCachingService.removeCachedLayout(owner, profile);
+        }
 
         // cache the relatively lightwieght userprofile for use in 
         // in layout PLF loading
